Add IndicatorObject.
This object implements IndicatorIface and is a wrapper built around Indicator.Object. Thanks to it the rest of Wingpanel will no longer depend on the Indicator API directly, but on IndicatorIface and IndicatorWidget. It also controls entry additions and removals. This was previously done in Wingpanel.Panel using association by hash map, which was not safe.